Add ai generated documentation
