+++ app/src/main/kotlin/dev/atick/compose/MainActivity.kt
@@ -37,7 +37,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import dagger.hilt.android.AndroidEntryPoint
-import dev.atick.bluetooth.common.models.BtState
+import dev.atick.bluetooth.common.model.BtState
 import dev.atick.bluetooth.common.utils.BluetoothUtils
 import dev.atick.compose.ui.JetpackApp
 import dev.atick.core.extensions.isAllPermissionsGranted
@@ -205,11 +205,17 @@ private fun shouldUseDarkTheme(
     }
 }
 
+/**
+ * Determines whether a user is logged in based on the provided [UiState].
+ *
+ * @param uiState The UI state representing the user data.
+ * @return `true` if the user is considered logged in; `false` otherwise.
+ */
 private fun isUserLoggedIn(uiState: UiState<UserData>): Boolean {
     return when (uiState) {
-        is UiState.Loading -> true
-        is UiState.Success -> uiState.data.id != "-1"
-        is UiState.Error -> false
+        is UiState.Loading -> true // User is considered logged in during loading (assuming ongoing session).
+        is UiState.Success -> uiState.data.id.isNotEmpty() // User is logged in if the data ID is not "-1".
+        is UiState.Error -> false // User is not logged in in case of an error.
     }
 }
