+++ app/src/main/kotlin/dev/atick/compose/data/profile/ProfileScreenData.kt
@@ -16,7 +16,13 @@
 
 package dev.atick.compose.data.profile
 
+/**
+ * Represents data for displaying a user's profile screen.
+ *
+ * @property name The user's name to be displayed on the profile screen.
+ * @property profilePictureUri The URI for the user's profile picture, or null if not available.
+ */
 data class ProfileScreenData(
-    val name: String = "",
+    val name: String = String(),
     val profilePictureUri: String? = null,
 )
