+++ app/src/main/kotlin/dev/atick/compose/repository/home/PostsRepositoryImpl.kt
@@ -20,8 +20,8 @@ import dev.atick.compose.data.home.UiPost
 import dev.atick.compose.data.home.mapToPostEntities
 import dev.atick.compose.data.home.mapToUiPost
 import dev.atick.compose.data.home.toUiPost
-import dev.atick.network.NetworkDataSource
-import dev.atick.storage.room.LocalDataSource
+import dev.atick.network.data.NetworkDataSource
+import dev.atick.storage.room.data.LocalDataSource
 import dev.atick.storage.room.model.PostEntity
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
