+++ auth/src/main/kotlin/dev/atick/auth/config/Config.kt
@@ -1,6 +1,14 @@
 package dev.atick.auth.config
 
+/**
+ * Configuration object containing constants and values used in the application.
+ */
 object Config {
+    /**
+     * The Web Client ID used for Google Sign-In authentication.
+     * This identifier is associated with your application and is used to authenticate with Google services.
+     * Ensure that it is properly configured and secured in the Google Developer Console.
+     */
     const val WEB_CLIENT_ID =
         "1052755869243-tb3s2g60ct3fji814vq59ht295946ai2.apps.googleusercontent.com"
 }
