+++ auth/src/main/kotlin/dev/atick/auth/di/DataSourceModule.kt
@@ -8,10 +8,19 @@ import dev.atick.auth.data.AuthDataSource
 import dev.atick.auth.data.AuthDataSourceImpl
 import javax.inject.Singleton
 
+/**
+ * Dagger Hilt module for providing data source dependencies.
+ */
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class DataSourceModule {
 
+    /**
+     * Binds the [AuthDataSourceImpl] implementation to the [AuthDataSource] interface.
+     *
+     * @param authDataSourceImpl The implementation of [AuthDataSource] to be bound.
+     * @return An instance of [AuthDataSource] for dependency injection.
+     */
     @Binds
     @Singleton
     abstract fun bindAuthDataSource(
