+++ auth/src/main/kotlin/dev/atick/auth/di/RepositoryModule.kt
@@ -8,15 +8,22 @@ import dev.atick.auth.repository.AuthRepository
 import dev.atick.auth.repository.AuthRepositoryImpl
 import javax.inject.Singleton
 
-
+/**
+ * Dagger Hilt module for providing repository dependencies.
+ */
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class RepositoryModule {
 
+    /**
+     * Binds the [AuthRepositoryImpl] implementation to the [AuthRepository] interface.
+     *
+     * @param authRepositoryImpl The implementation of [AuthRepository] to be bound.
+     * @return An instance of [AuthRepository] for dependency injection.
+     */
     @Binds
     @Singleton
     abstract fun bindAuthRepository(
         authRepositoryImpl: AuthRepositoryImpl,
     ): AuthRepository
-
 }
