+++ auth/src/main/kotlin/dev/atick/auth/ui/AuthViewModel.kt
@@ -4,7 +4,7 @@ import android.content.Intent
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import dev.atick.auth.model.login.AuthScreenData
+import dev.atick.auth.model.AuthScreenData
 import dev.atick.auth.repository.AuthRepository
 import dev.atick.core.extensions.isEmailValid
 import dev.atick.core.extensions.isPasswordValid
