--- auth/src/main/kotlin/dev/atick/auth/ui/SignUpScreen.kt
@@ -1,8 +1,0 @@
-package dev.atick.auth.ui
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun SignUpScreen() {
-
-}
