+++ auth/src/main/kotlin/dev/atick/auth/ui/signin/SignInScreen.kt
@@ -25,9 +25,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalFocusManager
@@ -38,7 +35,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import dev.atick.auth.R
-import dev.atick.auth.model.login.AuthScreenData
+import dev.atick.auth.model.AuthScreenData
 import dev.atick.auth.ui.AuthViewModel
 import dev.atick.core.ui.component.JetpackButton
 import dev.atick.core.ui.component.JetpackOutlinedButton
