+++ bluetooth/common/src/main/kotlin/dev/atick/bluetooth/common/data/BluetoothDataSource.kt
@@ -16,7 +16,7 @@
 
 package dev.atick.bluetooth.common.data
 
-import dev.atick.bluetooth.common.models.BtMessage
+import dev.atick.bluetooth.common.model.BtMessage
 import kotlinx.coroutines.flow.StateFlow
 
 /**
