+++ bluetooth/common/src/main/kotlin/dev/atick/bluetooth/common/manager/BluetoothManager.kt
@@ -16,7 +16,7 @@
 
 package dev.atick.bluetooth.common.manager
 
-import dev.atick.bluetooth.common.models.BtDevice
+import dev.atick.bluetooth.common.model.BtDevice
 import kotlinx.coroutines.flow.StateFlow
 
 /**
