+++ bluetooth/common/src/main/kotlin/dev/atick/bluetooth/common/model/BtMessage.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.atick.bluetooth.common.models
+package dev.atick.bluetooth.common.model
 
 import java.util.Date
