+++ network/src/main/kotlin/dev/atick/network/data/NetworkDataSourceImpl.kt
@@ -17,7 +17,6 @@
 package dev.atick.network.data
 
 import dev.atick.core.di.IoDispatcher
-import dev.atick.network.NetworkDataSource
 import dev.atick.network.api.JetpackRestApi
 import dev.atick.network.model.NetworkPost
 import kotlinx.coroutines.CoroutineDispatcher
