+++ storage/preferences/src/main/kotlin/dev/atick/storage/preferences/model/UserData.kt
@@ -24,7 +24,7 @@ import kotlinx.serialization.Serializable
  * This data class is used to store information about a user, including their ID, name, profile picture URI string,
  * preferred theme brand, dark theme configuration, and dynamic color preference.
  *
- * @property id The unique identifier for the user. Defaults to "-1" if not provided.
+ * @property id The unique identifier for the user. Defaults to empty if not provided.
  * @property name The name of the user. Defaults to "No Name" if not provided.
  * @property profilePictureUriString The URI string for the user's profile picture, if available. Defaults to `null` if not provided.
  * @property themeBrand The preferred theme brand for the user. Defaults to [ThemeBrand.DEFAULT].
@@ -33,7 +33,7 @@ import kotlinx.serialization.Serializable
  */
 @Serializable
 data class UserData(
-    val id: String = "-1",
+    val id: String = String(),
     val name: String = "No Name",
     val profilePictureUriString: String? = null,
     val themeBrand: ThemeBrand = ThemeBrand.DEFAULT,
