+++ storage/room/src/main/kotlin/dev/atick/storage/room/data/LocalDataSourceImpl.kt
@@ -17,7 +17,6 @@
 package dev.atick.storage.room.data
 
 import dev.atick.core.di.IoDispatcher
-import dev.atick.storage.room.LocalDataSource
 import dev.atick.storage.room.model.PostEntity
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
